refactor(conversation): initial protocol for conversation
