+++ openhands/sdk/conversation/impl/__init__.py
@@ -1,0 +1,4 @@
+from openhands.sdk.conversation.impl.local_conversation import LocalConversation
+
+
+__all__ = ["LocalConversation"]
